Moved verification code to respective controllers
