+++ apps/block_scout_web/lib/block_scout_web/controllers/address_contract_controller.ex
@@ -5,7 +5,7 @@ defmodule BlockScoutWeb.AddressContractController do
   require Logger
 
   alias BlockScoutWeb.AccessHelpers
-  alias BlockScoutWeb.AddressContractVerificationController, as: VerificationController
+  alias BlockScoutWeb.AddressContractVerificationViaJsonController, as: VerificationController
   alias Explorer.{Chain, Market}
   alias Explorer.ExchangeRates.Token
   alias Indexer.Fetcher.CoinBalanceOnDemand
