+++ apps/block_scout_web/lib/block_scout_web/notifier.ex
@@ -6,8 +6,6 @@ defmodule BlockScoutWeb.Notifier do
   alias Absinthe.Subscription
 
   alias BlockScoutWeb.{
-    AddressContractVerificationViaFlattenedCodeView,
-    AddressContractVerificationViaJsonView,
     ChangesetView,
     Endpoint
   }
