+++ apps/block_scout_web/lib/block_scout_web/templates/address_contract_verification_via_json/new.html.eex
@@ -23,7 +23,7 @@
         <label for="smart_contract_metadata_json"><%= gettext("Sources and Metadata JSON") %></label>
           <div class="center-column">
             <%= form_for @changeset,
-            address_contract_verification_path(@conn, :create),
+            address_verify_contract_via_json_path(@conn, :create, @address_hash),
             [class: "dropzone-1", name: "smart_contract[file]", style: "display: flex; margin: 0 auto;", id: "metadata-json-dropzone"],
             fn f -> %>
               <div style="text-align: center;">
