+++ apps/block_scout_web/lib/block_scout_web/templates/address_contract_verification_vyper/new.html.eex
@@ -13,7 +13,7 @@
       <h1 class="smart-contract-title"><%= gettext "New Vyper Smart Contract Verification" %></h1>
 
       <%= form_for @changeset,
-          address_contract_verification_path(@conn, :create),
+          address_verify_vyper_contract_path(@conn, :create, @address_hash),
           [],
           fn f -> %>
